+++ frontend/src/pythonTypes.ts
@@ -28,6 +28,18 @@ export interface Search {
     search_name: null | string;
 }
 
+export interface MinimalSession {
+    session_id: string;
+    folder_hash: string;
+    best_candidate: MinimalBestCandidateInfo;
+}
+
+export interface MinimalBestCandidateInfo {
+    data_source: string;
+    distance: number;
+    duplicates: Array<number>;
+}
+
 export interface LibraryStats {
     libraryPath: string;
     items: number;